就地重命名的目标页改存 id，防改到别的文件
问题：重命名输入框停留期间关掉它前面的一页，再打开一个文件补进空出来的下标
→ 一次回车（或点 ✓）会把**用户从没选过的文件**改名：fs::rename 落在别的页上，
页路径、写盘比对戳、最近文件与光标记忆全部跟着迁移。更早一步的可见症状是标签
条上那台输入框「长」在新打开的页签里。

原因：renaming_tab 存的是裸下标、提交时按下标现取页；而 close_tab_now 只给
close_tab_confirm 做了左移平移（P145），批量关闭（上一条提交）和这条都不在覆盖
范围内。

改法：renaming_tab 改存 TabId，提交时按 id 解析当前位置；解析不到 = 该页已被
关掉，输入态就地作废并清掉输入串。视图侧「哪一页渲染输入框」的判定同步改成按
tab.id 比，与状态字段同一个口径。

验证：全量 769 绿；clippy --all-targets -D warnings 与 cargo doc -D warnings
exit=0。新增 rename_commit_targets_the_page_it_started_on_not_the_index 用三个
真实临时文件跑完整链路（被改名的必须是当初那一页、c.txt 不得被碰、路径迁移落在
对的页上）。两处分别证过敏感：整条链退回旧形状 → 编码断言红（left Some(1) vs
right Some(2)）；只把提交端的 id 解析退回按下标 → 行为断言红（「改名应落在 b
上」）。另把五处 assert_eq!(renaming_tab, Some(0)) 从「等于下标」改成「等于该页
id」——它们此前测的是编码而不是行为。

+++ crates/app/src/state.rs
@@ -81,8 +81,10 @@ pub(crate) struct Editpad {
     /// 按标签顺序排列。None/空 = 提示条不可见。Esc（BarsDismissed）/
     /// 逐个裁决/全部忽略即清。
     pub(crate) external_change: Option<Vec<usize>>,
-    /// P55：就地重命名的目标页下标；Some = 标签条上该页显示为输入框。
-    pub(crate) renaming_tab: Option<usize>,
+    /// P55：就地重命名的目标页 **id**（Some = 标签条上该页显示为输入框）。
+    /// P214：曾是裸下标——输入框停留期间关掉前面任意一页，一次回车就会把
+    /// **另一个文件**改名（见 `commit_tab_rename`）。
+    pub(crate) renaming_tab: Option<u64>,
     /// P55：就地重命名的输入内容（预填当前文件名，纯 UI 态）。
     pub(crate) rename_input: String,
     /// P67：状态栏「编码」弹出菜单可见。
